+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Added `OK` and `Cancel` buttons to the `Advanced settings` window.
 
+## Fixed
+- Allow a reference `xlsx` file to contain `None` entries in the `mass_modifier` column.
+
 
 ## [0.2.2]
 ### Fixed